+++ cranelift/codegen/src/isa/legacy/riscv/enc_tables.rs
